#19 revert mod.ts and move existing code to old folder
