+++ src/old/classes/Log.ts
