+++ src/old/classes/RollupConfiguration.ts
