+++ src/old/templates/MANIFEST.MF
