+++ src/old/templates/manifest.json
